jx_export_shell -> jx_print_shell
