+++ dttools/src/jx_print.c
@@ -248,3 +248,28 @@ char * jx_print_string( struct jx *j )
 	buffer_free(&buffer);
 	return str;
 }
+
+static char * unquoted_string( struct jx *j )
+{
+	char *str;
+	if(j->type==JX_STRING) {
+		str = strdup(j->u.string_value);
+	} else {
+		str = jx_print_string(j);
+	}
+	return str;
+}
+
+/*
+Export a JX object as environment variables in bash format.
+*/
+
+void jx_print_shell( struct jx *j, FILE *stream )
+{
+	struct jx_pair *p;
+	for(p=j->u.pairs;p;p=p->next) {
+		char *str = unquoted_string(p->value);
+		fprintf(stream,"export %s=%s\n",p->key->u.string_value,str);
+		free(str);
+	}
+}
